+++ inst/include/bgenstream.h
@@ -34,43 +34,6 @@ namespace gds {
 /** Construct view of BGEN file using index to subset variants
  * based on region or variant id
  * @param filename path to BGEN file
- * @param index_filename path to index for BGEN file
- * @param gr `GenomicRanges` of intervals
- * @param rsids vector<string> of variant ids
- */ 
-static genfile::bgen::View::UniquePtr construct_view(
-	const string & filename,
-	const string & index_filename,
-	const GenomicRanges & gr,
-	const vector<string> & rsids = vector<string>()) {
-
-	if( ! filesystem::exists( filename ) ){
-		throw runtime_error("File does not exist: " + filename);
-	}
-	if( ! filesystem::exists( index_filename ) ){
-		throw runtime_error("File does not exist: " + index_filename);
-	}
-
-	using namespace genfile::bgen ;
-
-	View::UniquePtr view = View::create( filename ) ;
-
-	if( gr.size() > 0){		
-		IndexQuery::UniquePtr query = IndexQuery::create( index_filename ) ;
-		for( int i = 0; i < gr.size(); i++ ) {
-			query->include_range( IndexQuery::GenomicRange( gr.get_chrom(i) , gr.get_start(i), gr.get_end(i) ) ) ;
-		}
-		query->include_rsids( rsids ) ;
-		query->initialise() ;
-		view->set_query( query ) ;
-	}
-	return view ;
-}
-
-
-/** Construct view of BGEN file using index to subset variants
- * based on region or variant id
- * @param filename path to BGEN file
  */ 
 static genfile::bgen::View::UniquePtr construct_view(
 	const string & filename) {
@@ -96,6 +59,37 @@ static IndexQuery::UniquePtr construct_query(const string &index_filename){
 
 
 
+/** Construct view of BGEN file using index to subset variants
+ * based on region or variant id
+ * @param filename path to BGEN file
+ * @param index_filename path to index for BGEN file
+ * @param gr `GenomicRanges` of intervals
+ * @param rsids vector<string> of variant ids
+ */ 
+static genfile::bgen::View::UniquePtr construct_view(
+	const string & filename,
+	const string & index_filename,
+	const GenomicRanges & gr) {
+	// const vector<string> & rsids = vector<string>()) {
+
+	// create view of BGEN file
+	View::UniquePtr view = construct_view( filename );
+
+	// process region queryies
+	if( gr.size() > 0){		
+		IndexQuery::UniquePtr query = construct_query(index_filename);
+		for( int i = 0; i < gr.size(); i++ ) {
+			query->include_range( IndexQuery::GenomicRange( gr.get_chrom(i) , gr.get_start(i), gr.get_end(i) ) ) ;
+		}
+		// query->include_rsids( rsids ) ;
+		query->initialise() ;
+		view->set_query( query ) ;
+	}
+
+	return view ;
+}
+
+
 
 /** bgenstream reads a BGEN into an matrix in chunks, storing variants in columns.  Applies filtering for specified samples and genome region. 
  * 
@@ -103,6 +97,8 @@ static IndexQuery::UniquePtr construct_query(const string &index_filename){
 class bgenstream : 
 	public GenomicDataStream {
 	public:
+
+	string filenameIdxGlobal;
 
 	bgenstream() {}
 
@@ -113,6 +109,7 @@ class bgenstream :
 		// Initialize view from just bgen file
 		view = construct_view( param.file ) ;
 
+		filenameIdxGlobal = param.file + ".bgi";
 		queryGlobal = construct_query( param.file + ".bgi" );
 
 		// apply region filters
@@ -155,7 +152,9 @@ class bgenstream :
 		
 		GenomicRanges gr( regions );
 
-		auto query = queryGlobal;
+		// auto query = queryGlobal;
+
+		auto query = construct_query( filenameIdxGlobal );
 
 		if( gr.size() > 0){	
 			for( int i = 0; i < gr.size(); i++ ) {
@@ -169,6 +168,9 @@ class bgenstream :
 
 		// number of variants after filtering 
 		n_variants_total = view->number_of_variants() ;
+
+		// set current position of index
+		variant_idx_start = 0;
 	}
 
 	/** Get number of columns in data matrix
@@ -270,7 +272,7 @@ class bgenstream :
 	vector<double> matDosage;
 	size_t max_entries_per_sample = 4;
 	int n_variants_total;	
-	int variant_idx_start = 0;
+	int variant_idx_start;
 
 	bool getNextChunk_helper(){	
 
@@ -307,7 +309,7 @@ class bgenstream :
 
 			// read variant information
 			view->read_variant( &SNPID, &rsid, &chromosome, &position, &alleles ) ;
-		
+
 			// store variant info
 			vInfo->addVariant(chromosome, position, rsid, alleles[0], alleles[1] );