+++ src/export.cpp
@@ -63,6 +63,7 @@ SEXP create_xptr(
 	Param param( file, region, samples, chunkSize, missingToMean);
 	param.setField(field);
  
+ 	// calls constructor for GenomicDataStream
 	Rcpp::XPtr<BoundDataStream> z( new BoundDataStream(param), true);
 
 	return z;
@@ -77,6 +78,22 @@ List getInfo(SEXP x){
 	return List::create(Named("streamType") = ptr->ptr->getStreamType(),
 						Named("nsamples") = ptr->ptr->n_samples());
 }
+
+
+// [[Rcpp::export]]
+SEXP setRegions_rcpp( SEXP x, const string &regionString){
+	Rcpp::XPtr<BoundDataStream> ptr(x);
+
+	vector<string> regions = splitRegionString( regionString );
+
+	ptr->ptr->setRegions( regions );
+
+	ptr->atEndOfStream = false;
+	ptr->featuresRead = 0;
+
+	return ptr;
+}
+
 
 
 // [[Rcpp::export]]